backend/src/requirements.txt
asyncpg==0.24.0
alembic==1.6.5
asgiref==3.4.1
click==8.0.1
databases==0.4.3
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pydantic==1.8.2
email-validator
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-multipart==0.0.5
PyJWT==2.1.0
passlib[bcrypt]==1.7.2
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
